requirements.txt
discord.py[voice]>=1.7.3
supabase>=0.2.8
pillow
openai
